+++ src/resource_browser/hints.rs
@@ -7,11 +7,21 @@ pub(crate) const HELP_HINTS: &[&str] = &[
     "0..9 sort",
     "Enter props",
 ];
+
+pub(crate) const HELP_HINTS_EVENTS: &[&str] = &[
+    "q quit",
+    "/ search",
+    "r resource",
+    "0..9 sort",
+    "Enter details",
+];
+
 const EXPAND_NETWORK: &str = "n network";
 const EXPAND_DATASTORE: &str = "d datastore";
 const EXPAND_HOST: &str = "h host";
 const EXPAND_VM: &str = "v vm";
 const EXPAND_TASK: &str = "t task";
+const EXPAND_EVENT: &str = "e events";
 const VM_ACTIONS: &str = "x actions";
 
 const CLUSTER_EXPAND_HINTS: &[&str] = &[
@@ -21,11 +31,17 @@ const CLUSTER_EXPAND_HINTS: &[&str] = &[
     EXPAND_VM,
     EXPAND_TASK,
 ];
-const HOST_EXPAND_HINTS: &[&str] = &[EXPAND_NETWORK, EXPAND_DATASTORE, EXPAND_VM, EXPAND_TASK];
-const DATASTORE_EXPAND_HINTS: &[&str] = &[EXPAND_HOST, EXPAND_VM, EXPAND_TASK];
+const HOST_EXPAND_HINTS: &[&str] = &[
+    EXPAND_NETWORK,
+    EXPAND_DATASTORE,
+    EXPAND_VM,
+    EXPAND_TASK,
+    EXPAND_EVENT,
+];
+const DATASTORE_EXPAND_HINTS: &[&str] = &[EXPAND_HOST, EXPAND_VM, EXPAND_TASK, EXPAND_EVENT];
 const NETWORK_EXPAND_HINTS: &[&str] = &[EXPAND_HOST, EXPAND_VM, EXPAND_TASK];
 
-const VM_EXPAND_HINTS: &[&str] = &[VM_ACTIONS, EXPAND_TASK];
+const VM_EXPAND_HINTS: &[&str] = &[VM_ACTIONS, EXPAND_TASK, EXPAND_EVENT];
 pub(crate) fn get_expand_hint(resource_type: ResourceType) -> &'static [&'static str] {
     match resource_type {
         ResourceType::Cluster => CLUSTER_EXPAND_HINTS,
@@ -33,6 +49,6 @@ pub(crate) fn get_expand_hint(resource_type: ResourceType) -> &'static [&'static
         ResourceType::Datastore => DATASTORE_EXPAND_HINTS,
         ResourceType::Network => NETWORK_EXPAND_HINTS,
         ResourceType::VirtualMachine => VM_EXPAND_HINTS,
-        _ => &[],
+        ResourceType::Task | ResourceType::Event => &[],
     }
 }