+++ src/resource_browser/mod.rs
@@ -1,15 +1,17 @@
 mod cluster;
 mod data_loaders;
 mod datastore;
+pub mod events;
 mod formatting;
 mod hints;
 mod host;
 mod indexed_cache;
 mod network;
+mod perf;
 mod resource_mgr;
 mod resource_table;
 mod tabular_data;
 mod task;
 mod vm;
-
+pub use events::{EventBrowserPayload, event_to_browser_object};
 pub use resource_mgr::{HistoryRecord, ResourceManager};